Moving files around to better reflect different projects
