+++ src/menus/menu.js
@@ -1,7 +1,22 @@
-import {StaticCharacter} from "../characters/static_character";
 import {ACTION_NONE} from "./menu_actions.js";
 import {FACING_DOWN, FACING_UP} from "../core/facing";
 import {wrapText} from "../core/text_helpers";
+
+import {Character} from "../core/character.js";
+
+class StaticCharacter extends Character {
+    constructor(initialX, initialY, symbol) {
+         super(initialX, initialY);
+         this.symbol = symbol;
+     }
+ 
+     getCharacter(col, row) {
+         if (this.getX() == col && this.getY() == row) {
+             return this.symbol;
+         }
+         return null;
+     }
+ }
 
 export class Menu {
     constructor(menuSpec, viewport, zPosition, spaceChar) {
