+++ src/treasure_hunt/characters/static_character.js
@@ -1,4 +1,4 @@
-import {Character} from "../core/character.js";
+import {Character} from "../../core/character.js";
 
 export class StaticCharacter extends Character {
     constructor(initialX, initialY, symbol) {
