+++ src/treasure_hunt/key_map.js
@@ -1,4 +1,4 @@
-import {FACING_DOWN, FACING_LEFT, FACING_RIGHT, FACING_UP} from "./core/facing.js";
+import {FACING_DOWN, FACING_LEFT, FACING_RIGHT, FACING_UP} from "../core/facing.js";
 
 export class KeyMap {
     constructor() {
