+++ src/treasure_hunt/map.js
